First prototype of a space discovery tool.
The general idea is to gather all the SiteInfo tiddlers
which have been tagged with pre-defined tags (see SPACE_TAGS)
and then present each space in a lump related to the tag. When
a space title is clicked, its SiteInfo is loaded for sake of
review, and an (inactive for now) "Include!" button is there
for use. From the SiteInfo it is easy to navigate to the space,
to see what it is about.

Bootstrap is used to make various bits dismissable (in case they
are in the way). SiteInfo is loaded into modal displays.

If something like this were to exist, then it would need to be availble
as an in the space doing the including, otherwise DropPrivs would
prevent the inclusion proces.

Issues:

* links in SiteInfo are not in correct context
* quite a few SiteInfo tiddlers are not relevant
* what are the correct tags to be "lumping"?
* It's still in just English
* It provides functionality but it doesn't provide fun.

However, despite the issues it at least provides something
concrete to complain about and fork and fiddle with, rather
than being confused and disoriented.